- added rudimentary support of obj-Files arch model works perfectly I hope - extended many classes with debugOutputs -> wxString ToString() - MainFrame creates a rootAssembly and every opened obj File is a child of rootassembly - find bug? -------------> inform

+++ src/GNRVertex.cpp
@@ -19,7 +19,7 @@ GNRVertex::GNRVertex()
 	this->m_z = 0.0;
 }
 
-GNRVertex::GNRVertex(double x, double y, double z)
+GNRVertex::GNRVertex(float x, float y, float z)
 {
 	this->m_x = x;
 	this->m_y = y;
